pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "black>=25.9.0",
    "isort>=6.0.1",
    "langchain[openai]>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langchain-openai>=0.3.33",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.6.7",
    "python-dotenv>=1.1.1",
    "pydantic>=2.11.9",
]
